add decode fuzz tests for block, tx and receipt
